Revert "Patch from DukePyrolator to replace all calls to match_wild() and match_wild_nocase() to use Anope::Match() instead. Also changes line-endings on wildcard.cpp to Unix-style, as well as fixes a small compile warning in language.c."
This reverts commit fcab9857f55567f10eaecbd6b26ee96f0f549d65.

This isn't a simple sed operation, the order of arguments changed, so this would break everything using it.

git-svn-id: http://anope.svn.sourceforge.net/svnroot/anope/trunk@2081 5417fbe8-f217-4b02-8779-1006273d7864

+++ src/language.c
@@ -130,7 +130,7 @@ static void load_lang(int index, const char *filename)
 		} else {
 			langtexts[index][i] = new char[len + 1];
 			fseek(f, pos, SEEK_SET);
-			if ((int32)fread(langtexts[index][i], 1, len, f) != len) {
+			if (fread(langtexts[index][i], 1, len, f) != len) {
 				alog("Failed to read string %d in language %d (%s)",
 					 i, index, filename);
 				while (--i >= 0) {
